refactor(editor): extract telemetry module
