+++ editor/js/editor-libs/codemirror-editor.js
@@ -23,7 +23,7 @@ import { parser as jsParser } from "@lezer/javascript";
 import { parser as htmlParser } from "@lezer/html";
 import { parser as cssParser } from "@lezer/css";
 
-import { postActionMessage } from "./events.js";
+import { postActionMessage } from "./telemetry.js";
 
 import "../../css/editor-libs/codemirror-override.css";
